+++ examples/resources/cloudtemple_compute_iaas_opensource_virtual_machine/import.sh
@@ -1,0 +1,4 @@
+#!/bin/bash
+
+# Import a virtual machine using its ID
+terraform import cloudtemple_compute_iaas_opensource_virtual_machine.example 12345678-1234-1234-1234-123456789abc